backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
Flask-Migrate==4.0.5
psycopg2-binary==2.9.9
boto3==1.34.0
werkzeug==3.0.1
requests==2.31.0
python-dotenv==1.0.0
PyJWT==2.8.0
gunicorn==21.2.0
anthropic==0.40.0
